+++ CHANGELOG.md
@@ -7,7 +7,10 @@
 -   Surface the `validation_interval_in_trees`,
     `keep_non_leaf_label_distribution` and 'random_seed' hyper-parameters.
 -   Add the `batch_size` argument in the `pd_dataframe_to_tf_dataset` utility.
--   Automatically determine the number of threads if `num_threads=None` .
+-   Automatically determine the number of threads if `num_threads=None`.
+-   Check that the training dataset is well configured for TF-DF e.g. no repeat
+    operation, has a large enough batch size, etc. The check can be disabled
+    with `check_dataset=False`.
 
 ### Fixes